requirements.txt
fastapi>=0.68.0
uvicorn==0.40.0
openai==2.16.0
python-dotenv==1.2.1
dashscope>=1.0.0
